Implement GET /excursions
