+++ backend/app.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const authRoutes = require("./routes/authRoutes");
-const tourRoutes = require("./routes/tourRoutes");
+const tourRoutes = require("./routes/excursionRoutes");
 // const exampleRoutes = require('./routes/exampleRoutes');
 
 const app = express();
