+++ backend/models/excursionModel.js
@@ -1,0 +1,14 @@
+const { sql } = require("../utils/postgres");
+
+exports.getAllExcursionsM = async () => {
+  const excursionsList = await sql`
+    SELECT 
+    tours.*, 
+    categories.name AS category_name
+    FROM tours
+    JOIN categories 
+    ON tours.category_id = categories.id;
+    `;
+
+  return excursionsList;
+};
